Refactor event components and auto-update event type
Moved event-related modules from 'services' to new 'components' subfolders for better modularity. Updated imports in cogs and handlers to reflect new structure. Added logic to automatically update the event type based on frequency changes in EventsService. Updated documentation to reflect new architecture and event type behavior.
