+++ components/choices/__init__.py
@@ -1,0 +1,1 @@
+# Componentes de choices
